:bug: bug: fixed bad formatter Data Class Json - change Product.kt price Float to Double

+++ core/src/main/kotlin/id/barakkastudio/core/data/datasource/local/db/entity/ProductEntity.kt
@@ -14,7 +14,7 @@ data class ProductEntity(
     val description: String? = null,
 
     @ColumnInfo(name = "price")
-    val price: Long? = null,
+    val price: Double? = null,
 
     @ColumnInfo(name = "thumbnail")
     val thumbnail: String? = null,
